관리자 코드 분리
